ai_bot_services/ai_service/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
groq
sentry-sdk
openai